Some input parameters like accountId and AssetId removed

+++ FinanceServicesApi/V1/Boundary/Responses/ResidentSummary/ResidentSummaryResponse.cs
@@ -10,7 +10,7 @@ namespace FinanceServicesApi.V1.Boundary.Responses.ResidentSummary
         /// <example>
         ///     3215.32
         /// </example>
-        public float? CurrentBalance { get; set; }
+        public decimal? CurrentBalance { get; set; }
 
         /// <summary>
         ///
